source/lambda/iot-dr-r53-health-check/requirements.txt
awsiotsdk
